requirements.txt
Flask==2.0.2
MarkupSafe==2.1.0
mysql-connector-python
 
